delete popup

+++ resources/views/profile.blade.php
@@ -23,7 +23,7 @@
                             <div class="img-thumbnail">
                                 <img alt="" src="http://localhost/IMGY/public/{{$image->url}}" width="250px" height="200px" />
                                 @if(Auth::user()->id == $user->id)
-                                    <a href="/IMGY/public/delete/{{$image->id}}"><button style="position: absolute; top: 5px; right: 23px;" class="btn btn-danger btn-xs"><span class="glyphicon glyphicon-trash"></span></button></a> 
+                                    <a onclick="return confirmAction()" href="/IMGY/public/delete/{{$image->id}}"><button style="position: absolute; top: 5px; right: 23px;" class="btn btn-danger btn-xs"><span class="glyphicon glyphicon-trash"></span></button></a> 
                                 @endif
                             </div>
                         </a>
@@ -34,5 +34,13 @@
 
         </div>
     </div>
+    <script>
+            
+            function confirmAction(){
+            var confirmed = confirm("Are you sure? This will remove image forever.");
+           return confirmed;
+            }
+
+         </script>
 @endsection
